+++ changelog.markdown
@@ -1,6 +1,7 @@
 ## 1.50.0.0
 
 * New task DeleteDirs allows to delete multiple directories.
+* New parameter for NuGet dependencies.
 
 ## 1.48.0.0